Clean up FilterAnalysis and add javadoc on StupidPool
