--- .idea/mavenProjectSettings.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="MavenProjectSettings">
-    <option name="testRunningSettings">
-      <MavenTestRunningSettings>
-        <option name="passArgLine" value="false" />  <!-- see https://github.com/apache/incubator-druid/pull/8526 -->
-      </MavenTestRunningSettings>
-    </option>
-  </component>
-</project>
