+++ processing/src/main/java/org/apache/druid/segment/QueryableIndexCursorSequenceBuilder.java
@@ -30,19 +30,14 @@ import org.apache.druid.java.util.common.guava.Sequence;
 import org.apache.druid.java.util.common.guava.Sequences;
 import org.apache.druid.java.util.common.io.Closer;
 import org.apache.druid.query.BaseQuery;
-import org.apache.druid.query.BitmapResultFactory;
-import org.apache.druid.query.DefaultBitmapResultFactory;
 import org.apache.druid.query.Query;
 import org.apache.druid.query.QueryMetrics;
 import org.apache.druid.query.filter.Filter;
 import org.apache.druid.query.monomorphicprocessing.RuntimeShapeInspector;
-import org.apache.druid.segment.column.BitmapColumnIndex;
 import org.apache.druid.segment.column.ColumnHolder;
 import org.apache.druid.segment.column.NumericColumn;
 import org.apache.druid.segment.data.Offset;
 import org.apache.druid.segment.data.ReadableOffset;
-import org.apache.druid.segment.filter.AndFilter;
-import org.apache.druid.segment.filter.Filters;
 import org.apache.druid.segment.historical.HistoricalCursor;
 import org.apache.druid.segment.vector.BitmapVectorOffset;
 import org.apache.druid.segment.vector.FilteredVectorOffset;
@@ -56,9 +51,6 @@ import org.joda.time.Interval;
 
 import javax.annotation.Nullable;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 
 public class QueryableIndexCursorSequenceBuilder
 {
@@ -107,10 +99,11 @@ public class QueryableIndexCursorSequenceBuilder
         columnCache
     );
 
-    final QueryableIndexStorageAdapter.FilterAnalysis filterAnalysis = analyzeFilter(
+    final FilterAnalysis filterAnalysis = FilterAnalysis.analyzeFilter(
         filter,
         bitmapIndexSelector,
-        metrics
+        metrics,
+        index.getNumRows()
     );
 
     final ImmutableBitmap filterBitmap = filterAnalysis.getPreFilterBitmap();
@@ -217,10 +210,11 @@ public class QueryableIndexCursorSequenceBuilder
         columnCache
     );
 
-    final QueryableIndexStorageAdapter.FilterAnalysis filterAnalysis = analyzeFilter(
+    final FilterAnalysis filterAnalysis = FilterAnalysis.analyzeFilter(
         filter,
         bitmapIndexSelector,
-        metrics
+        metrics,
+        index.getNumRows()
     );
 
     final ImmutableBitmap filterBitmap = filterAnalysis.getPreFilterBitmap();
@@ -340,94 +334,6 @@ public class QueryableIndexCursorSequenceBuilder
 
     // Not found.
     return endIndex;
-  }
-
-  @SuppressWarnings("rawtypes")
-  private QueryableIndexStorageAdapter.FilterAnalysis analyzeFilter(
-      @Nullable final Filter filter,
-      ColumnSelectorColumnIndexSelector indexSelector,
-      @Nullable QueryMetrics queryMetrics
-  )
-  {
-    final int totalRows = index.getNumRows();
-
-    /*
-     * Filters can be applied in two stages:
-     * pre-filtering: Use bitmap indexes to prune the set of rows to be scanned.
-     * post-filtering: Iterate through rows and apply the filter to the row values
-     *
-     * The pre-filter and post-filter step have an implicit AND relationship. (i.e., final rows are those that
-     * were not pruned AND those that matched the filter during row scanning)
-     *
-     * An AND filter can have its subfilters partitioned across the two steps. The subfilters that can be
-     * processed entirely with bitmap indexes (subfilter returns non-null value for getBitmapColumnIndex)
-     * will be moved to the pre-filtering stage.
-     *
-     * Any subfilters that cannot be processed entirely with bitmap indexes will be moved to the post-filtering stage.
-     */
-    final List<Filter> preFilters;
-    final List<Filter> postFilters = new ArrayList<>();
-    int preFilteredRows = totalRows;
-    if (filter == null) {
-      preFilters = Collections.emptyList();
-    } else {
-      preFilters = new ArrayList<>();
-
-      if (filter instanceof AndFilter) {
-        // If we get an AndFilter, we can split the subfilters across both filtering stages
-        for (Filter subfilter : ((AndFilter) filter).getFilters()) {
-
-          final BitmapColumnIndex columnIndex = subfilter.getBitmapColumnIndex(indexSelector);
-
-          if (columnIndex == null) {
-            postFilters.add(subfilter);
-          } else {
-            preFilters.add(subfilter);
-            if (!columnIndex.getIndexCapabilities().isExact()) {
-              postFilters.add(subfilter);
-            }
-          }
-        }
-      } else {
-        // If we get an OrFilter or a single filter, handle the filter in one stage
-        final BitmapColumnIndex columnIndex = filter.getBitmapColumnIndex(indexSelector);
-        if (columnIndex == null) {
-          postFilters.add(filter);
-        } else {
-          preFilters.add(filter);
-          if (!columnIndex.getIndexCapabilities().isExact()) {
-            postFilters.add(filter);
-          }
-        }
-      }
-    }
-
-    final ImmutableBitmap preFilterBitmap;
-    if (preFilters.isEmpty()) {
-      preFilterBitmap = null;
-    } else {
-      if (queryMetrics != null) {
-        BitmapResultFactory<?> bitmapResultFactory =
-            queryMetrics.makeBitmapResultFactory(indexSelector.getBitmapFactory());
-        long bitmapConstructionStartNs = System.nanoTime();
-        // Use AndFilter.getBitmapIndex to intersect the preFilters to get its short-circuiting behavior.
-        preFilterBitmap = AndFilter.getBitmapIndex(indexSelector, bitmapResultFactory, preFilters);
-        preFilteredRows = preFilterBitmap.size();
-        queryMetrics.reportBitmapConstructionTime(System.nanoTime() - bitmapConstructionStartNs);
-      } else {
-        BitmapResultFactory<?> bitmapResultFactory = new DefaultBitmapResultFactory(indexSelector.getBitmapFactory());
-        preFilterBitmap = AndFilter.getBitmapIndex(indexSelector, bitmapResultFactory, preFilters);
-      }
-    }
-
-    if (queryMetrics != null) {
-      queryMetrics.preFilters(new ArrayList<>(preFilters));
-      queryMetrics.postFilters(postFilters);
-      queryMetrics.reportSegmentRows(totalRows);
-      queryMetrics.reportPreFilteredRows(preFilteredRows);
-    }
-
-    return new QueryableIndexStorageAdapter.FilterAnalysis(preFilterBitmap, Filters.maybeAnd(postFilters).orElse(null));
   }
 
   private static class QueryableIndexVectorCursor implements VectorCursor
